+++ bot/buffers.js
