+++ bot/log.js
